+++ lib/ui_chat_basic.dart
@@ -11,6 +11,7 @@ import 'tts_service.dart';
 import 'widget_help_dialog.dart';
 import 'utility.dart';
 import 'math_keyboard.dart';
+import 'tpSpeechText.dart';
 
 class chat {
   int p; //0:自分 1:相手
@@ -1169,46 +1170,6 @@ class ChatBubbleTriangle extends CustomPainter {
   bool shouldRepaint(CustomPainter oldDelegate) => false;
 }
 
-//読み上げ用テキスト変換関数
-//読み上げ用テキスト変換関数
-String toSpeechText(String inputText) {
-  // 数学系
-  inputText = inputText.replaceAllMapped(RegExp(r'([a-zA-Z0-9])\^([a-zA-Z0-9])'), (m) => '${m[1]}${m[2]}乗');
-  inputText = inputText.replaceAllMapped(RegExp(r'([a-zA-Z0-9]+)\/([a-zA-Z0-9]+)'), (m) => '${m[2]}分の${m[1]}');
-  inputText = inputText.replaceAll(r'=', 'イコール');
-  inputText = inputText.replaceAll(r'-', 'マイナス');
-
-  // ~形
-  inputText = inputText.replaceAll('現在形', '現在けい');
-  inputText = inputText.replaceAll('過去形', '過去けい');
-  inputText = inputText.replaceAll('進行形', '進行けい');
-
-  // www → 笑
-  inputText = inputText.replaceAll(RegExp(r'\b(w{2,})\b', caseSensitive: false), '笑');
-
-  // 記号
-  inputText = inputText.replaceAll(RegExp(r'[.…]{2,}'), ''); // 3点
-  inputText = inputText.replaceAll(RegExp(r'"'), ''); // アスタリスク
-  inputText = inputText.replaceAll(RegExp(r':'), ''); // コロン
-
-  // 絵文字の削除
-  final symbolEmojiRegex = RegExp(
-    r'''
-  (?:[\u2300-\u23FF]|
-     [\u2600-\u26FF]|
-     [\u2700-\u27BF]|
-     \u{1F300}-\u{1F5FF}|
-     \u{1F600}-\u{1F64F}|
-     \u{1F680}-\u{1F6FF}|
-     \u{1F900}-\u{1F9FF}
-    )
-  ''',
-    unicode: true,
-  );
-  inputText = inputText.replaceAll(symbolEmojiRegex, '');
-
-  return inputText;
-}
 class MessageDialog extends StatelessWidget {
   final String feedbackMessage;
   final String modelPath;
